+++ ErsatzTV.Application/Scheduling/Queries/GetDecoTemplateItemsHandler.cs
@@ -10,11 +10,16 @@ public class GetDecoTemplateItemsHandler(IDbContextFactory<TvContext> dbContextF
     {
         await using TvContext dbContext = await dbContextFactory.CreateDbContextAsync(cancellationToken);
 
+        // drop items that are invalid
         return await dbContext.DecoTemplateItems
             .AsNoTracking()
             .Filter(i => i.DecoTemplateId == request.DecoTemplateId)
             .Include(i => i.Deco)
             .ToListAsync(cancellationToken)
-            .Map(items => items.Map(Mapper.ProjectToViewModel).ToList());
+            .Map(
+                items => items
+                    .Map(Mapper.ProjectToViewModel)
+                    .Filter(i => i.StartTime < i.EndTime || i.EndTime.TimeOfDay == TimeSpan.Zero)
+                    .ToList());
     }
 }