+++ ErsatzTV.Application/Scheduling/Queries/GetTemplateItemsHandler.cs
@@ -10,11 +10,16 @@ public class GetTemplateItemsHandler(IDbContextFactory<TvContext> dbContextFacto
     {
         await using TvContext dbContext = await dbContextFactory.CreateDbContextAsync(cancellationToken);
 
+        // drop items that are invalid
         return await dbContext.TemplateItems
             .AsNoTracking()
             .Filter(i => i.TemplateId == request.TemplateId)
             .Include(i => i.Block)
             .ToListAsync(cancellationToken)
-            .Map(items => items.Map(Mapper.ProjectToViewModel).ToList());
+            .Map(
+                items => items
+                    .Map(Mapper.ProjectToViewModel)
+                    .Filter(i => i.StartTime < i.EndTime || i.EndTime.TimeOfDay == TimeSpan.Zero)
+                    .ToList());
     }
 }